requirements.txt
fastapi==0.103.2
orjson==3.9.7
pydantic==2.4.2
pydantic-settings==2.0.3
uvicorn==0.23.2
uvloop==0.17.0
alembic==1.12.0
asyncpg==0.28.0
pytest==7.4.2
pytest-env==1.0.1
pytest-mock==3.11.1
python-dotenv==1.0.0
SQLAlchemy==2.0.21
SQLAlchemy-Utils==0.41.1
starlette==0.31.1
pyshorteners==1.0.1
factory-boy==3.3.0
aiosqlite==0.19.0
Faker==18.6.0
httpx==0.24.0
